requirements.txt
openai~=1.58.1
requests~=2.32.3
python-dotenv~=1.0.1